Move priorities toggle into routing settings
