+++ frontend/src/features/settings/components/appearance-settings.test.tsx
@@ -1,12 +1,11 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { beforeEach, describe, expect, it, vi } from "vitest";
+import { beforeEach, describe, expect, it } from "vitest";
 
 import { AppearanceSettings } from "@/features/settings/components/appearance-settings";
 import { useAccountQuotaDisplayStore } from "@/hooks/use-account-quota-display";
 import { useThemeStore } from "@/hooks/use-theme";
 import { useTimeFormatStore } from "@/hooks/use-time-format";
-import type { DashboardSettings } from "@/features/settings/schemas";
 
 function installLocalStorageMock() {
   const storage = new Map<string, string>();
@@ -27,32 +26,17 @@ function installLocalStorageMock() {
   });
 }
 
-const BASE_SETTINGS: DashboardSettings = {
-  stickyThreadsEnabled: true,
-  upstreamStreamTransport: "default",
-  preferEarlierResetAccounts: true,
-  prioritiesEnabled: false,
-  routingStrategy: "capacity_weighted",
-  openaiCacheAffinityMaxAgeSeconds: 1800,
-  dashboardSessionTtlSeconds: 43200,
-  importWithoutOverwrite: true,
-  totpRequiredOnLogin: false,
-  totpConfigured: false,
-  apiKeyAuthEnabled: false,
-};
-
 describe("AppearanceSettings", () => {
   beforeEach(() => {
     installLocalStorageMock();
     useThemeStore.setState({ preference: "light", theme: "light", initialized: true });
     useTimeFormatStore.setState({ timeFormat: "12h" });
     useAccountQuotaDisplayStore.setState({ quotaDisplay: "both" });
-  });
-
+});
   it("exposes selected state for the time-format toggle", async () => {
     const user = userEvent.setup();
 
-    render(<AppearanceSettings settings={BASE_SETTINGS} busy={false} onSave={vi.fn().mockResolvedValue(undefined)} />);
+    render(<AppearanceSettings />);
 
     const button12h = screen.getByRole("button", { name: /12h/i });
     const button24h = screen.getByRole("button", { name: /24h/i });
@@ -70,7 +54,7 @@ describe("AppearanceSettings", () => {
   it("exposes selected state for the account quota toggle", async () => {
     const user = userEvent.setup();
 
-    render(<AppearanceSettings settings={BASE_SETTINGS} busy={false} onSave={vi.fn().mockResolvedValue(undefined)} />);
+    render(<AppearanceSettings />);
 
     const button5h = screen.getByRole("button", { name: "5H" });
     const buttonWeekly = screen.getByRole("button", { name: "W" });
@@ -91,28 +75,4 @@ describe("AppearanceSettings", () => {
     expect(useAccountQuotaDisplayStore.getState().quotaDisplay).toBe("weekly");
   });
 
-  it("saves the priorities toggle", async () => {
-    const user = userEvent.setup();
-    const onSave = vi.fn().mockResolvedValue(undefined);
-
-    render(<AppearanceSettings settings={BASE_SETTINGS} busy={false} onSave={onSave} />);
-
-    const prioritiesSwitch = screen.getByRole("switch");
-    expect(prioritiesSwitch).toHaveAttribute("aria-checked", "false");
-
-    await user.click(prioritiesSwitch);
-
-    expect(onSave).toHaveBeenCalledWith({
-      stickyThreadsEnabled: true,
-      upstreamStreamTransport: "default",
-      preferEarlierResetAccounts: true,
-      prioritiesEnabled: true,
-      routingStrategy: "capacity_weighted",
-      openaiCacheAffinityMaxAgeSeconds: 1800,
-      dashboardSessionTtlSeconds: 43200,
-      importWithoutOverwrite: true,
-      totpRequiredOnLogin: false,
-      apiKeyAuthEnabled: false,
-    });
-  });
 });
